pyproject.toml
[tool.poetry]
name = "png2svg"
version = "0.1.0"
description = ""
authors = ["Athon Millane <athon@sadana.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
svgwrite = "^1.4.3"
pillow = "^11.1.0"
numpy = "^2.2.4"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
